+++ source/extensions/omni.isaac.benchmarks/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [0.18.2] - 2023-10-09
+### Changed
+- changed carter bench mark tests from Carter V2 to Carter V1
 ## [0.18.1] - 2023-08-21
 
 ### Changed